issue #117 refactor: conflict resolve
