+++ src/constants/iconConstants.js
@@ -1,9 +1,33 @@
 import CloseIcon from "@/assets/icon/ic-close.svg";
+import CrownIcon from "@/assets/icon/ic-crown.svg";
+import CommentIcon from "@/assets/icon/ic-feed-comment.svg";
+import HeartIcon from "@/assets/icon/ic-feed-heart.svg";
+import OptionThreeDotIcon from "@/assets/icon/ic-feed-option.svg";
 import GoogleLogoIcon from "@/assets/icon/ic-google-logo.svg";
+import AddIcon from "@/assets/icon/ic-group-add.svg";
+import IntroductionEditIcon from "@/assets/icon/ic-introduction-edit.svg";
+import LeftArrowIcon from "@/assets/icon/ic-left-arrow.svg";
 import NotificationDropdownBubbleIcon from "@/assets/icon/ic-notification-dropdown.svg";
 import ProfileDropdownBubbleIcon from "@/assets/icon/ic-profile-dropdown.svg";
+import RequestCheckIcon from "@/assets/icon/ic-request-check.svg";
+import RequestCloseIcon from "@/assets/icon/ic-request-close.svg";
+import RightArrowIcon from "@/assets/icon/ic-right-arrow.svg";
+import SearchIcon from "@/assets/icon/ic-search.svg";
+import SecretIcon from "@/assets/icon/ic-secret.svg";
 
 export {
+	SearchIcon,
+	IntroductionEditIcon,
+	LeftArrowIcon,
+	RightArrowIcon,
+	CrownIcon,
+	CommentIcon,
+	HeartIcon,
+	OptionThreeDotIcon,
+	SecretIcon,
+	RequestCheckIcon,
+	RequestCloseIcon,
+	AddIcon,
 	CloseIcon,
 	GoogleLogoIcon,
 	NotificationDropdownBubbleIcon,
